+++ annotations/src/main/java/com/karuslabs/annotations/Monad.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2018 Karus Labs.
+ * Copyright 2020 Karus Labs.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,13 +21,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.karuslabs.commons.util.concurrent.locks;
+package com.karuslabs.annotations;
 
+import java.lang.annotation.*;
 
-public interface Holdable {
-    
-    Mutex hold();
-    
-    Mutex holdInterruptibly() throws InterruptedException;
-    
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.SOURCE;
+   
+@Documented
+@Retention(SOURCE)
+@Target(TYPE)
+public @interface Monad {
+
 }