+++ commons/src/main/java/com/karuslabs/commons/command/aot/annotations/Infer.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2018 Karus Labs.
+ * Copyright 2020 Karus Labs.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,17 +21,21 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.karuslabs.annotations;
+package com.karuslabs.commons.command.aot.annotations;
 
 import java.lang.annotation.*;
 
-import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.ElementType.*;
 import static java.lang.annotation.RetentionPolicy.SOURCE;
 
 
 @Documented
 @Retention(SOURCE)
-@Target(TYPE)
-public @interface ValueType {
+@Target(PARAMETER)
+public @interface Infer {
+
+    public static String INFERRED_ARGUMENT = "${inferred argument}";
+    
+    public String value() default INFERRED_ARGUMENT;
     
 }