+++ commons/src/test/resources/aot/generation/Commands.java
@@ -11,7 +11,7 @@ import org.bukkit.command.CommandSender;
 
 
 /**
- * This file was generated at test-time using Chimera 4.8.0
+ * This file was generated at test-time using Chimera 4.9.0-SNAPSHOT
  */
 public class Commands {